+++ src/main/kotlin/csense/idea/base/bll/kotlin/KtClassOrObject.kt
@@ -1,6 +1,6 @@
 package csense.idea.base.bll.kotlin
 
-import com.intellij.psi.PsiElement
+import org.jetbrains.kotlin.asJava.classes.KtLightClassForSourceDeclaration
 import org.jetbrains.kotlin.idea.references.resolveMainReferenceToDescriptors
 import org.jetbrains.kotlin.js.resolve.diagnostics.findPsi
 import org.jetbrains.kotlin.lexer.KtTokens
@@ -24,6 +24,9 @@ fun KtClassOrObject.isAnonymous(): Boolean = name == null
 
 fun KtClassOrObject.getAllFunctions(): List<KtNamedFunction> = collectDescendantsOfType()
 
+/**
+ *
+ */
 val KtClassOrObject.superClass: KtClassOrObject?
     get() {
         val superTypes = superTypeListEntries
@@ -31,10 +34,9 @@ val KtClassOrObject.superClass: KtClassOrObject?
             return null
         }
         superTypes.forEach {
-            val realClass = it.typeAsUserType?.referenceExpression?.resolveMainReferenceToDescriptors()
-                ?.firstOrNull()?.containingDeclaration?.findPsi() as? KtClassOrObject
-            if (realClass != null) {
-                return realClass
+            when (val resolved = it.typeReference?.resolve()) {
+                is KtLightClassForSourceDeclaration -> return resolved.kotlinOrigin
+                is KtClassOrObject -> return resolved
             }
         }
         return null